requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
bitarray==3.0.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
crcmod==1.7
cryptography==43.0.3
Cython==3.0.11
datasets==2.18.0
decorator==5.1.1
dill==0.3.8
einops==0.8.0
eva-decord==0.6.1
exceptiongroup==1.2.2
fairseq==0.12.2
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
gast==0.6.0
h11==0.14.0
huggingface-hub==0.26.2
hydra-core==1.0.7
idna==3.10
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
MarkupSafe==3.0.2
modelscope==1.20.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.0.6
open_clip_torch==2.29.0
opencv-python==4.10.0.84
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.0.0
propcache==0.2.0
pyarrow==18.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.41.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.20.3
tomli==2.1.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
unicodedata2==15.1.0
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
xxhash==3.5.0
yapf==0.43.0
yarl==1.17.2
zhconv==1.4.3
